Updates:
AllDessertsView: Replaced LazyVStack with a regular VStack. LazyVStack requires Caching the images. This will be implemented as time allows.

DessertRowView: Replaced AsyncImage with QuickImageView for maintainability.

IngredientListView: Added checks to make sure the value of the ingredients were not equal to a blank string or nil. Refactored the view into four smaller fives that fill out the screen better.

QuickImage: Changed the Image of the .failed state to an x mark. Changed the let property to the value of an optional URL to avoid forced unwrapping.

DessertDetailsView: Added a section to provide Links  for the source and Youtube if present. Added a navigation title. Removed redundant VStack. Removed function that attended to unwrap URL unnecessarily.

DessertDetailsModel: made all properties except the meal name, ID, and thumbnail optional.

AllDessertsDataService: Added print statements if the URL guard statement fails. Subscribed top the dataTaskPuiblisher on the main thread.