Updates: Dependency Injection.
Renamed AllDessertsDataService to DessertsDataService

DessertDetailsViewModel: Instead of creating a new instance of DessertDataService, the DessertDetailsViewModel now has a dependency of DessertDataService. The data service is initialized in the AllDessertsViewModel then passed into the DessertDetailsView when upon init it is passed into the DessertDetailsViewModel.

DessertDetailsModel: Created a static placeholder instance of the model.