Update: Safety Improvements.
If a user taps on the navigation link before the image has loaded the image will not be displayed when they return to the AllDessertsView. To prevent this from happening i used a .disabled modifier on the view to allow the image to finish loading. I added an overlay to the navigation label to indicate that the view is still loading. This is a temporary solution, i will continue work on the problem